add habit and listView habit completed
