+++ lib/Reducer/AppState.dart
@@ -1,5 +1,4 @@
-import 'package:add_habit_demo_3/habit.dart';
-import 'package:add_habit_demo_3/habit_list.dart';
+import 'package:add_habit_demo_3/AddHabit/habit.dart';
 import 'package:redux/redux.dart';
 
 class AppState {
