+++ lib/main.dart
@@ -1,10 +1,8 @@
-import 'package:add_habit_demo_3/add_habit_form.dart';
-import 'package:add_habit_demo_3/app_state.dart';
+import 'package:add_habit_demo_3/AddHabit/AddHabitForm.dart';
+import 'package:add_habit_demo_3/Reducer/AppState.dart';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
-
-import 'habit_list.dart';
 
 void main() {
   runApp(MyApp());
@@ -34,8 +32,7 @@ class HabitTrackerScreen extends StatelessWidget {
       appBar: AppBar(
         title: const Text('Habit Tracker'),
       ),
-      body: HabitList(),
-      floatingActionButton: AddHabitForm(),
+      body: AddHabitForm(),
     );
   }
 }
